add init module
